refactor (service): common heartbeat
cdsctl admin repositories|vcs|hooks status
