+++ engine/hatchery/kubernetes/kubernetes.go
@@ -61,13 +61,13 @@ func (h *HatcheryKubernetes) ApplyConfiguration(cfg interface{}) error {
 		}
 	}
 
-	// s.Client = cdsclient.NewService(s.Cfg.API.HTTP.URL, 60*time.Second)
-	// s.API = s.Cfg.API.HTTP.URL
-	// s.Name = s.Cfg.Name
-	// s.HTTPURL = s.Cfg.URL
-	// s.Token = s.Cfg.API.Token
-	// s.Type = services.TypeHatchery
-	// s.MaxHeartbeatFailures = s.Cfg.API.MaxHeartbeatFailures
+	// h.Client = cdsclient.NewService(h.Config.API.HTTP.URL, 60*time.Second)
+	// h.API = h.Config.API.HTTP.URL
+	// h.Name = h.Config.Name
+	// h.HTTPURL = h.Config.URL
+	// h.Token = h.Config.API.Token
+	// h.Type = services.TypeHatchery
+	// h.MaxHeartbeatFailures = h.Config.API.MaxHeartbeatFailures
 
 	return nil
 }
