+++ engine/hatchery/kubernetes/types.go
@@ -5,10 +5,10 @@ import (
 	"sync"
 	"time"
 
+	"github.com/ovh/cds/engine/service"
 	"github.com/ovh/cds/sdk"
 	"github.com/ovh/cds/sdk/cdsclient"
 	"github.com/ovh/cds/sdk/hatchery"
-	"github.com/ovh/cds/sdk/service"
 
 	"k8s.io/client-go/kubernetes"
 )
