+++ engine/hatchery/marathon/types.go
@@ -3,10 +3,10 @@ package marathon
 import (
 	"github.com/gambol99/go-marathon"
 
+	"github.com/ovh/cds/engine/service"
 	"github.com/ovh/cds/sdk"
 	"github.com/ovh/cds/sdk/cdsclient"
 	"github.com/ovh/cds/sdk/hatchery"
-	"github.com/ovh/cds/sdk/service"
 )
 
 // HatcheryConfiguration is the configuration for hatchery
