+++ engine/hatchery/swarm/types.go
@@ -3,10 +3,10 @@ package swarm
 import (
 	docker "github.com/docker/docker/client"
 
+	"github.com/ovh/cds/engine/service"
 	"github.com/ovh/cds/sdk"
 	"github.com/ovh/cds/sdk/cdsclient"
 	"github.com/ovh/cds/sdk/hatchery"
-	"github.com/ovh/cds/sdk/service"
 )
 
 // HatcheryConfiguration is the configuration for hatchery
